before fix for cluster_l sequence load, all other utils and algorithms working ok

+++ website/cdvrp.py
@@ -46,11 +46,11 @@ class aco:
             '''
         )
 
-        max_id_set = pd.read_sql_query(points_query, self.con)
-        if max_id_set.iloc[0, 0] == None:
-            max_val = 0
-        else:
-            max_val = int.from_bytes(max_id_set.iloc[0, 0], 'little') + 1
+        # max_id_set = pd.read_sql_query(points_query, self.con)
+        # if max_id_set.iloc[0, 0] == None:
+        #     max_val = 0
+        # else:
+        #     max_val = int.from_bytes(max_id_set.iloc[0, 0], 'little') + 1
 
         cluster_l = []
         cluster_counter = 0
@@ -60,24 +60,29 @@ class aco:
             if status == False:
                 sequence_n = 0
             if item == self.origin and status == False:
-                cluster_l.append([max_val, self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
+                # cluster_l.append([max_val, self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
+                cluster_l.append([self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
                 sequence_n += 1
                 status = True
             elif item != self.origin and status == True:
-                cluster_l.append([max_val, self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
+                # cluster_l.append([max_val, self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
+                cluster_l.append([self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
                 sequence_n += 1
             elif item != self.origin and status == False:
                 status = True
-                cluster_l.append([max_val, self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
+                # cluster_l.append([max_val, self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
+                cluster_l.append([self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
                 sequence_n += 1
             elif item == self.origin and status == True:
                 status = False
-                cluster_l.append([max_val, self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
+                # cluster_l.append([max_val, self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
+                cluster_l.append([self.dvrp_id, cluster_counter, 'tractor_' + str(cluster_counter), item, sequence_n])
                 cluster_counter += 1
             
-        cluster_l = [x for x in cluster_l if x[4] != self.origin]
+        cluster_l = [x for x in cluster_l if x[3] != self.origin]
 
-        self.cur.executemany('insert into dvrp_set (id_set, dvrp_id, cluster_id, cluster_name, point, sequence) values (?, ?, ?, ?, ?, ?)', cluster_l)
+        # self.cur.executemany('insert into dvrp_set (id_set, dvrp_id, cluster_id, cluster_name, point, sequence) values (?, ?, ?, ?, ?, ?)', cluster_l)
+        self.cur.executemany('insert into dvrp_set (dvrp_id, cluster_id, cluster_name, point, sequence) values (?, ?, ?, ?, ?)', cluster_l)
         self.cur.execute('insert into dvrp_origin (dvrp_id, dvrp_origin) values (?, ?)', [self.dvrp_id, self.origin])
         self.con.commit()
 
@@ -91,6 +96,22 @@ class aco:
 
         self.con = sqlite3.connect(self.db_path)
         self.cur = self.con.cursor()
+
+        dvrp_id_query = (
+            '''
+            select exists (
+                select 1
+                from dvrp_origin
+                where dvrp_id = ?
+            ) as dvrp_id_exists
+            ;
+            '''
+        )
+
+        res_exists = pd.read_sql_query(dvrp_id_query, self.con, params=[self.dvrp_id])
+        if res_exists.iloc[0, 0] == 1:
+            print('dvrp_id already exists in the database.')
+            return
 
         self.cur.execute('create temporary table if not exists temp_items (item Text)')
         self.cur.execute('delete from temp_items')
@@ -210,6 +231,8 @@ class aco:
 
         self.load_best_path_id_p()
         self.con.close()
+
+        return 'dvrp_set solved and loaded'
 
 
 functions = {
